create shared richTextContent schema
