+++ src/components/RichText/types.ts
@@ -1,26 +1,23 @@
+import { RichTextContent } from "types/components";
+
 type ReadOnly = {
-  readOnly: true;
-  //map editable attr to never
+	readOnly: true;
+	//map editable attr to never
 } & Partial<{ [key in keyof Omit<Editable, "readOnly">]?: never }>;
 
 export type Editable = {
-  readOnly?: never;
-  onChange(content: RichTextContent): void;
-  onError(error: string): void;
-  placeHolder?: string;
-  charLimit?: number;
-  disabled?: boolean;
-  invalid?: boolean;
+	readOnly?: never;
+	onChange(content: RichTextContent): void;
+	onError(error: string): void;
+	placeHolder?: string;
+	charLimit?: number;
+	disabled?: boolean;
+	invalid?: boolean;
 };
 
 export type EditorClasses = { container?: string; charCounter?: string };
 
 export type Props = (ReadOnly | Editable) & {
-  content: RichTextContent;
-  classes?: EditorClasses;
-};
-
-export type RichTextContent = {
-  value: string;
-  length: number;
+	content: RichTextContent;
+	classes?: EditorClasses;
 };
