+++ src/types/components.ts
@@ -6,13 +6,18 @@ export type OptionType<V> = { label: string; value: V };
 
 //dropzone
 export type FileDropzoneAsset = {
-  previews: FileObject[]; //from previous submission
-  files: File[]; //new files
+	previews: FileObject[]; //from previous submission
+	files: File[]; //new files
 };
 
 //country selector
 export type Country = {
-  name: string;
-  flag: string;
-  code: string;
+	name: string;
+	flag: string;
+	code: string;
+};
+
+export type RichTextContent = {
+	value: string;
+	length: number;
 };
